DI: add test case for interface
